backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
aiofiles>=23.2.1
pymongo[srv]==4.7.2
certifi==2024.2.2
bcrypt==4.1.3
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
tensorflow==2.16.1
h5py>=3.9.0
python-dotenv==1.0.1
httpx>=0.27.0
opencv-python-headless==4.9.0.80
